% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BET_MSE.R
\name{BET_MSE}
\alias{BET_MSE}
\title{Runs the BET MSE framework for the specified number of iterations}
\usage{
BET_MSE(pdir, sdir, hsnum, hcrnum, scnnum, itr, Bthr, Blim, Tmin, pl)
}
\arguments{
\item{pdir}{OM directory path}

\item{hsnum}{number characterizing the harvest strategy being run}

\item{hcrnum}{number characterizing the harvest control rule being run}

\item{scnnum}{number characterizing the uncertainty scenario being run}

\item{itr}{the iteration that one wants to run}

\item{Bthr}{the fraction of unfished biomass the threshold reference point refers to}

\item{Blim}{the fraction of unfished biomass the limit reference point refers to}

\item{Tmin}{specifies the fraction of the minimum catch at the LRP}

\item{pl}{specifies the level of probability of SSB>LRP required by the hcr
Note that the F target is specified in the forecast file. The fishing intensity is measured as the spawning potential ratio (SPR).
SPR is the calculated as the ratio of the equilibrium reproductive output per recruit that would occur
with the current year's F intensities and biology, to the equilibrium reproductive output per recruit
that would occur with the current year's biology and no fishing.}

\item{FBtrh}{the fraction of Btrh to consider for the threshold reference point}
}
\value{
a data frame of output for performance statistics
}
\description{
Runs the BET MSE framework for the specified number of iterations
}
\author{
D.Tommasi
}
