% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_OM.R
\name{Bootstrap_OM}
\alias{Bootstrap_OM}
\title{Make projection based on simulated R devs and HCR}
\usage{
Bootstrap_OM(dir_istep, istep, dir_OM, Mcycle, seed, plot = NA)
}
\arguments{
\item{dir_istep}{the directory of step i}

\item{dir_OM}{the directory of OM}

\item{Mcycle}{the number of years within a management cycle}

\item{seed}{the seed for bootstrap}

\item{plot}{what r4ss tabs to plot}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
