% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{Clean}
\alias{Clean}
\title{Clean unused folders to save space}
\usage{
Clean(pdir, sdir, HS, HCR, OM, itrnum, nquarters, Mcycle)
}
\arguments{
\item{pdir}{parent directory path}

\item{sdir}{benchmark assessment model directory path}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{the name of the operating model}

\item{itrnum}{iteration number}

\item{nquarters}{total number of new quarters to be simulated in the MSE}

\item{Mcycle}{the number of years within a management cycle}
}
\description{
Clean unused folders to save space
}
\author{
Haikun Xu
}
