% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation_EM_risk.R
\name{Estimation_EM_risk}
\alias{Estimation_EM_risk}
\title{Make projection based on simulated R devs and HCR}
\usage{
Estimation_EM_risk(
  dir_istep,
  dir_EM_previous,
  dir_OM_Boot,
  R0,
  Mcycle,
  dat_name,
  ctl_name,
  ss_name,
  plot = FALSE,
  include_LF = TRUE
)
}
\arguments{
\item{dir_istep}{the directory of step i}

\item{dir_EM_previous}{the directory of the EM in the previous step}

\item{dir_OM_Boot}{the directory of the OM bootstrap in the current step}

\item{R0}{true R0}

\item{Mcycle}{the number of years within a management cycle}

\item{dat_name}{name of the SS data file}

\item{ctl_name}{name of the SS control file}

\item{ss_name}{name of the SS exe file}

\item{plot}{whether to plot EM's results}

\item{include_LF}{whether to fit the EM to bootstrapped LF data}

\item{from_par}{whether to start the EM from the par file}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
