% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_OM.R
\name{Extract_OM}
\alias{Extract_OM}
\title{Extract and save the final OM's output}
\usage{
Extract_OM(dir_OM_Final, startquarter, plot = FALSE)
}
\description{
Extract and save the final OM's output
}
\author{
Haikun Xu
}
