% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_OM.R
\name{Final_OM}
\alias{Final_OM}
\title{Make the final OM for extracting MSE results}
\usage{
Final_OM(dir_itr, istep, dir_OM, Mcycle, endquarter)
}
\arguments{
\item{dir_itr}{the directory of iteration itr}

\item{istep}{step number}

\item{dir_OM}{the directory of OM}

\item{Mcycle}{the number of years within a management cycle}
}
\description{
Make the final OM for extracting MSE results
}
\author{
Haikun Xu
}
