% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR_2.R
\name{HCR_2}
\alias{HCR_2}
\title{Runs the staff's HCR algorithm}
\usage{
HCR_2(dir_EM, istep, CurrentClosure)
}
\arguments{
\item{dir_EM}{the directory of the EM used to compute the current SBR_d for harvest control rule}

\item{CurrentClosure}{the current number of closure days}
}
\description{
Runs the staff's HCR algorithm
}
\author{
Haikun Xu
}
