% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize_OM.R
\name{Initialize_OM}
\alias{Initialize_OM}
\title{Initialize the OM based on the benchmark assessment model}
\usage{
Initialize_OM(pdir, sdir, HS, HCR, OM, dat_name, ctl_name, ss_name, clean)
}
\arguments{
\item{pdir}{parent directory path}

\item{sdir}{the directory of the assessment models with F30\% reference points}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{name of the reference model}

\item{dat_name}{name of the SS data file}

\item{ctl_name}{name of the SS control file}

\item{ss_name}{name of the SS exe file}
}
\description{
Initialize the OM based on the benchmark assessment model
}
\author{
Haikun Xu
}
