% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_OM.R
\name{Projection_OM}
\alias{Projection_OM}
\title{Make projection based on simulated R devs and HCR}
\usage{
Projection_OM(
  pdir,
  HS,
  HCR,
  OM,
  itr,
  istep,
  Fratio,
  dir_OM_previous,
  dir_EM_previous,
  R_devs,
  n_extra_R,
  Mcycle,
  dat_name,
  ctl_name,
  ss_name,
  plot = FALSE
)
}
\arguments{
\item{pdir}{parent directory path}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{the name of the operating model}

\item{itr}{iteration number}

\item{istep}{step number}

\item{dir_OM_previous}{the directory of the OM in the previous step}

\item{dir_EM_previous}{the directory of the EM in the previous step}

\item{R_devs}{simulated R devs}

\item{n_extra_R}{the number of recruitment devs after the main R and before the forecast R}

\item{Mcycle}{the number of years within a management cycle}

\item{dat_name}{name of the SS data file}

\item{ctl_name}{name of the SS control file}

\item{ss_name}{name of the SS exe file}

\item{plot}{whether to plot OM's results}

\item{Fscale}{input F scaler}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
