% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_OM.R
\name{Update_OM}
\alias{Update_OM}
\title{Update OM's data file with "true" data}
\usage{
Update_OM(
  istep,
  dir_OM,
  dir_OM_Boot,
  dir_OM_Final,
  Mcycle,
  endquarter,
  dat_name,
  ss_name,
  ctl_name
)
}
\arguments{
\item{dir_OM}{the directory of OM}

\item{dir_OM_Boot}{the directory of the OM bootstrap in the current step}

\item{Mcycle}{the number of years within a management cycle}

\item{dat_name}{name of the SS data file}
}
\description{
Update OM's data file with "true" data
}
\author{
Haikun Xu
}
