% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_OM_SMSY.R
\name{Update_OM_SMSY}
\alias{Update_OM_SMSY}
\title{Update OM's data file with "true" data}
\usage{
Update_OM_SMSY(istep, dir_OM_Final, dir_OM_MSY, dat_name, ss_name, ctl_name)
}
\arguments{
\item{dat_name}{name of the SS data file}

\item{dir_OM}{the directory of OM}

\item{dir_OM_Boot}{the directory of the OM bootstrap in the current step}

\item{Mcycle}{the number of years within a management cycle}
}
\description{
Update OM's data file with "true" data
}
\author{
Haikun Xu
}
