% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BET_MSE.R
\name{BET_MSE}
\alias{BET_MSE}
\title{Runs the BET MSE framework for the specified number of iterations}
\usage{
BET_MSE(
  pdir,
  HS,
  HCR,
  OM,
  itrnum,
  nquarters,
  Mcycle,
  n_extra_R,
  startquarter,
  endquarter,
  EM_comp_fleet,
  dat_name,
  ctl_name,
  ss_name,
  clean = FALSE,
  plot = FALSE,
  MSY = FALSE
)
}
\arguments{
\item{pdir}{parent directory path}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{the name of the operating model}

\item{itrnum}{iteration number}

\item{nquarters}{total number of new quarters to be simulated in the MSE}

\item{Mcycle}{the number of years within a management cycle}

\item{n_extra_R}{the number of recruitment devs after the main R and before the forecast R}

\item{startquarter}{the first quarter of the assessment model}

\item{endquarter}{the last quarter of the assessment model}

\item{EM_comp_fleet}{The fleets with comp data in the EM}

\item{dat_name}{name of the SS data file}

\item{ctl_name}{name of the SS control file}

\item{ss_name}{name of the SS exe file}

\item{clean}{if TRUE, all intermediate folders of the MSE simulation will be deleted to save storage space}
}
\description{
Runs the BET MSE framework for the specified number of iterations
}
\author{
Haikun Xu
}
