% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BET_MSE.R
\name{BET_MSE}
\alias{BET_MSE}
\title{Runs the BET MSE framework for the specified number of iterations}
\usage{
BET_MSE(pdir, sdir, HS, HCR, OM, itrnum, nquarters, Mcycle, n_extra_R)
}
\arguments{
\item{pdir}{parent directory path}

\item{sdir}{benchmark assessment model directory path}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{the name of the operating model}

\item{itrnum}{iteration number}

\item{nquarters}{total number of new quarters to be simulated in the MSE}

\item{Mcycle}{the number of years within a management cycle}

\item{n_extra_R}{the number of recruitment devs after the main R and before the forecast R}
}
\description{
Runs the BET MSE framework for the specified number of iterations
}
\author{
Haikun Xu
}
