% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation_EM.R
\name{Estimationn_EM}
\alias{Estimationn_EM}
\title{Make projection based on simulated R devs and HCR}
\usage{
Estimationn_EM(
  dir_istep,
  R0,
  dir_OM_previous,
  dir_EM_previous,
  dir_OM_Boot,
  Mcycle,
  EM_comp_fleet,
  plot = FALSE
)
}
\arguments{
\item{dir_istep}{the directory of step i}

\item{R0}{true R0}

\item{dir_OM_previous}{the directory of the OM in the previous step}

\item{dir_EM_previous}{the directory of the EM in the previous step}

\item{dir_OM_Boot}{the directory of the OM bootstrap in the current step}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
