% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR_others.R
\name{HCR_others}
\alias{HCR_others}
\title{Runs the staff's HCR algorithm}
\usage{
HCR_others(dir_EM, istep, CurrentClosure, Scontrol, Fscaler, Sscaler)
}
\arguments{
\item{dir_EM}{the directory of the EM used to compute the current SBR_d for harvest control rule}

\item{CurrentClosure}{the current number of closure days}
}
\description{
Runs the staff's HCR algorithm
}
\author{
Haikun Xu
}
