% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize_OM.R
\name{Initialize_OM}
\alias{Initialize_OM}
\title{Initialize the OM based on the benchmark assessment model}
\usage{
Initialize_OM(pdir, sdir, HS, HCR, OM)
}
\arguments{
\item{pdir}{parent directory path}

\item{sdir}{the directory of the assessment models with F30\% reference points}

\item{OM}{name of the reference model}
}
\description{
Initialize the OM based on the benchmark assessment model
}
\author{
Haikun Xu
}
