% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_OM.R
\name{Projection_OM}
\alias{Projection_OM}
\title{Make projection based on simulated R devs and HCR}
\usage{
Projection_OM(
  pdir,
  HS,
  HCR,
  OM,
  itr,
  istep,
  Fscale,
  dir_OM_previous,
  dir_EM_previous,
  R_devs,
  n_extra_R,
  Mcycle,
  plot = NA
)
}
\arguments{
\item{pdir}{parent directory path}

\item{HS}{the name of the harvest strategy}

\item{HCR}{the name of the harvest control rule}

\item{OM}{the name of the operating model}

\item{itr}{iteration number}

\item{istep}{step number}

\item{Fscale}{input F scaler}

\item{dir_OM_previous}{the directory of the OM in the previous step}

\item{dir_EM_previous}{the directory of the EM in the previous step}

\item{R_devs}{simulated R devs}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
