% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_OM.R
\name{Update_OM}
\alias{Update_OM}
\title{Make projection based on simulated R devs and HCR}
\usage{
Update_OM(dir_OM, dir_OM_Boot, Mcycle, EM_comp_fleet)
}
\arguments{
\item{dir_OM_Boot}{the directory of the OM bootstrap in the current step}

\item{dir_istep}{the directory of step i}

\item{R0}{true R0}

\item{dir_EM_previous}{the directory of the EM in the previous step}
}
\description{
Make projection based on simulated R devs and HCR
}
\author{
Haikun Xu
}
